extract heatmap drawing logic to hook
